chat giữa client và server

+++ src/main/java/DAO/SessionDAOImpl.java
@@ -17,7 +17,7 @@ public class SessionDAOImpl extends BaseDAO implements ISessionDAO {
         preparedStatement.setTime(3, new java.sql.Time(session.getStartTime().getTime()));
         preparedStatement.setInt(4, session.getTotalTime());
         preparedStatement.setInt(5, session.getUsedTime());
-        preparedStatement.setInt(6, session.getUsedCost());
+        preparedStatement.setDouble(6, session.getUsedCost());
         preparedStatement.setInt(7, session.getServiceCost());
         preparedStatement.setDouble(8, session.getPrepaidAmount());
         preparedStatement.executeUpdate();
@@ -37,7 +37,7 @@ public class SessionDAOImpl extends BaseDAO implements ISessionDAO {
         preparedStatement.setTimestamp(3, new java.sql.Timestamp(session.getStartTime().getTime()));
         preparedStatement.setInt(4, session.getTotalTime());
         preparedStatement.setInt(5, session.getUsedTime());
-        preparedStatement.setInt(6, session.getUsedCost());
+        preparedStatement.setDouble(6, session.getUsedCost());
         preparedStatement.setInt(7, session.getServiceCost());
         preparedStatement.setDouble(8, session.getPrepaidAmount());
         preparedStatement.setInt(9, session.getId());
